fixup! Bug #38938: add layer for default UCR variables
Move import of cri to top: While it created an import loop, Python is
able to handle it and does this internally:

    sys.modules["univention.config_registry"] = {}
    sys.modules["univention.config_registry_info"] = {}
    ...
    sys.modules["univention.config_registry"].update(__import__("univention.config_registry"))
    ...
    sys.modules["univention.config_registry_info"].update(__import__("univention.config_registry_info"))

So univention.config_registry can do an "import univention.config_registry_info"
and already gets a reference to the first empty module definition for
univention.config_registry_info, which only will be fully usable after the
import is complete. Only after that you can reference objects from the
imported modules.